refactor: remove cargo workspaces.
`scripts/lib.rs` will not be that important for this project, so having the yew
app being the full project makes the configuration of the plugin a little bit
easier.

However, now the build script will just be there to have an easier way to get
the required files to the compile extention.

I still need to generate the files with `trunk`, but testing the website trough
trunk is easier than using web-ext anyways; gotta do it later; then, I gonna run
it on CI.