refactor: use `svg`s instead of `png`s.
If I wanna implement a dark theme I will need to be able to change the fill
color of the icons, so I decided to search a new set of icons that are svgs from
the get go.

refactor: remove download and edit button from layout.

I completely forgot that a `delete` button would be needed ( Idk how I missed
that ), so I decided to put the `download`, `edit` and `delete` actions inside a
new extra menu `...`. That should declutter the UI a little bit, but I will need
to implement another 'hamburger menu'; should be fairly easy right?